+++ Wms12m.Presentation/Areas/ToDo/Controllers/ProjectFormController.cs
@@ -15,6 +15,7 @@ namespace Wms12m.Presentation.Areas.ToDo.Controllers
         /// </summary>
         public ActionResult Index()
         {
+            if (CheckPerm(Perms.TodoProje, PermTypes.Reading) == false) return Redirect("/");
             ViewBag.MusteriID = new SelectList(db.Musteris.OrderBy(m => m.Unvan).ToList(), "ID", "Unvan ");
             ViewBag.Sorumlu = new SelectList(Persons.GetList(), "Kod", "AdSoyad");
             return View(new ProjeForm());
@@ -24,8 +25,8 @@ namespace Wms12m.Presentation.Areas.ToDo.Controllers
         /// </summary>
         public PartialViewResult List()
         {
+            if (CheckPerm(Perms.TodoProje, PermTypes.Reading) == false) return null;
             var projeForms = db.ProjeForms.Include(p => p.Musteri).Include(p => p.ProjeForm2);
-            // return PartialView(projeForms.ToList());
             return PartialView(projeForms.Where(a => a.PID == null).ToList());
         }
         /// <summary>
@@ -47,6 +48,7 @@ namespace Wms12m.Presentation.Areas.ToDo.Controllers
         [HttpPost, ValidateAntiForgeryToken]
         public JsonResult Save(ProjeForm projeForm)
         {
+            if (CheckPerm(Perms.TodoProje, PermTypes.Writing) == false) return Json(new Result(false, "Yetkiniz yok"), JsonRequestBehavior.AllowGet);
             if (ModelState.IsValid)
             {
                 if (projeForm.ID == 0)
@@ -86,6 +88,7 @@ namespace Wms12m.Presentation.Areas.ToDo.Controllers
         /// </summary>
         public JsonResult Delete(string Id)
         {
+            if (CheckPerm(Perms.TodoProje, PermTypes.Deleting) == false) return Json(new Result(false, "Yetkiniz yok"), JsonRequestBehavior.AllowGet);
             ProjeForm projeform = db.ProjeForms.Find(Id.ToInt32());
             db.ProjeForms.Remove(projeform);
             try
@@ -101,8 +104,9 @@ namespace Wms12m.Presentation.Areas.ToDo.Controllers
         /// <summary>
         /// formlar
         /// </summary>
-        public ActionResult FormIndex()
+        public PartialViewResult FormIndex()
         {
+            if (CheckPerm(Perms.TodoProje, PermTypes.Reading) == false) return null;
             var id = Url.RequestContext.RouteData.Values["id"];
             var ID = id.ToInt32();
             ProjeForm projeForm = db.ProjeForms.Find(ID);
@@ -120,6 +124,7 @@ namespace Wms12m.Presentation.Areas.ToDo.Controllers
         /// </summary>
         public PartialViewResult FormList()
         {
+            if (CheckPerm(Perms.TodoProje, PermTypes.Reading) == false) return null;
             var id = Url.RequestContext.RouteData.Values["id"];
             var ID = id.ToInt32();
             ViewBag.id = ID;
@@ -141,6 +146,7 @@ namespace Wms12m.Presentation.Areas.ToDo.Controllers
         [HttpPost, ValidateAntiForgeryToken]
         public JsonResult FormSave(ProjeForm projeForm)
         {
+            if (CheckPerm(Perms.TodoProje, PermTypes.Writing) == false) return Json(new Result(false, "Yetkiniz yok"), JsonRequestBehavior.AllowGet);
             if (ModelState.IsValid)
             {
                 if (projeForm.ID == 0)
